refactor(migration): separate what a rehearsal proves from what it drives
The smoke actions — register a name on v1, migrate it, register and renew
on v2, and the assertions each must satisfy — are the operations a real
user would take, used to show a phase did what it claimed rather than to
make the migration happen. They now live apart from the phases they check.

Getting them there needed the pieces both sides share to have a home
first, or the smoke module would have had to import from `migration.ts`,
which imports it. So the shared vocabulary — label ids, registration
durations, receipt waits — moved to the plumbing module, and the
transport-level pieces the wallet needs — the wallet client itself, the
gas buffer, time control — moved to the RPC module.

`migration.ts` drops from 9,923 to 9,066 lines.
